+++ docs/CHANGELOG.md
@@ -27,7 +27,7 @@ Version 1.3.0 (Bakeapple): Not yet released
 - Default `gibbing` to disabled for dedicated servers
 
 [@natarii](https://github.com/natarii)
-- Implement FFLink client functionality
+- Implement FFLink client functionality in launcher
 
 ### Bug fixes
 [@GooberRF](https://github.com/GooberRF)